Move background process code to main/worker
