+++ src/main/worker/__tests__/deviceService-test.js
