+++ src/main/worker/__tests__/settings-test.js
