+++ src/main/worker/deviceService.js
