+++ src/main/worker/settings.js
